Cifras de poblacion oficiales por comuna

Contiene los registros de poblacion desagregados por comuna, edad y sexo binario.

censo_2017_comunas

Format

Un data frame con 14.532 observaciones y cuatro columnas.

codigo_comuna

vector de texto

edad

vector categorico

sexo

vector categorico

poblacion

vector numerico

Source

Elaboracion propia con base en datos de Instituto Nacional de Estadisticas (INE).

Examples

censo_2017_comunas

#> # A tibble: 14,532 x 4
#>    codigo_comuna edad      sexo   poblacion
#>    <chr>         <fct>     <fct>      <dbl>
#>  1 01101         0 a 4     hombre      7186
#>  2 01101         10 a 14   hombre      6410
#>  3 01101         100 o mas hombre        15
#>  4 01101         15 a 19   hombre      6597
#>  5 01101         20 a 24   hombre      7574
#>  6 01101         25 a 29   hombre      8896
#>  7 01101         30 a 34   hombre      8180
#>  8 01101         35 a 39   hombre      7393
#>  9 01101         40 a 44   hombre      6586
#> 10 01101         45 a 49   hombre      5696
#> # ... with 14,522 more rows