Cifras de poblacion oficiales por zona

Contiene los registros de poblacion desagregados por zona y edad.

censo_2017_zonas

Format

Un data frame con 64.212 observaciones y tres columnas.

geocodigo

vector de texto

edad

vector categorico

poblacion

vector numerico

Source

Elaboracion propia con base en datos de Instituto Nacional de Estadisticas (INE).

Examples

censo_2017_zonas

#> # A tibble: 64,212 × 3
#>    geocodigo   edad  poblacion
#>    <chr>       <fct>     <dbl>
#>  1 01101011001 0 a 5       189
#>  2 01101011002 0 a 5       101
#>  3 01101021001 0 a 5        74
#>  4 01101021002 0 a 5         4
#>  5 01101021003 0 a 5       234
#>  6 01101021004 0 a 5       252
#>  7 01101021005 0 a 5       490
#>  8 01101031001 0 a 5       234
#>  9 01101031002 0 a 5       277
#> 10 01101031003 0 a 5       445
#> # … with 64,202 more rows