Codigos oficiales de comunas, provincias y regiones

Contiene los codigos oficiales de SUBDERE. Los nombres de las divisiones administrativas fueron convertidos a ASCII para evitar problemas de codificacion.

codigos_territoriales

Format

Un data frame con 346 observaciones y seis columnas.

codigo_comuna

vector de texto

nombre_comuna

vector de texto

codigo_provincia

vector de texto

nombre_provincia

vector de texto

codigo_region

vector de texto

nombre_region

vector de texto

Source

Elaboracion propia con base en datos de Subsecretaria de Desarrollo Regional (SUBDERE).

Examples

codigos_territoriales

#> # A tibble: 346 x 6
#>    codigo_comuna nombre_comuna codigo_provincia nombre_provincia codigo_region
#>    <chr>         <chr>         <chr>            <chr>            <chr>        
#>  1 01101         Iquique       011              Iquique          01           
#>  2 01107         Alto Hospicio 011              Iquique          01           
#>  3 01401         Pozo Almonte  014              Tamarugal        01           
#>  4 01402         Camina        014              Tamarugal        01           
#>  5 01403         Colchane      014              Tamarugal        01           
#>  6 01404         Huara         014              Tamarugal        01           
#>  7 01405         Pica          014              Tamarugal        01           
#>  8 02101         Antofagasta   021              Antofagasta      02           
#>  9 02102         Mejillones    021              Antofagasta      02           
#> 10 02103         Sierra Gorda  021              Antofagasta      02           
#> # … with 336 more rows, and 1 more variable: nombre_region <chr>