Codigos oficiales de comunas, provincias y regiones con 15 y 16 regiones

Contiene la conversion de los codigos oficiales de SUBDERE de 15 regiones a 16 regiones tras la creacion de la region de Niuble. Los nombres de las comunas fueron convertidos a ASCII para evitar problemas de codificacion.

codigos_territoriales_16r

Format

Un data frame con 21 observaciones y tres columnas.

codigo_comuna

vector de texto

codigo_comuna+16r

vector de texto

nombre_comuna

vector de texto

Source

Elaboracion propia con base en datos de Subsecretaria de Desarrollo Regional (SUBDERE) y CENSO 2017 (INE).

Examples

codigos_territoriales_16r

#> # A tibble: 21 × 3
#>    codigo_comuna codigo_comuna_16r nombre_comuna
#>    <chr>         <chr>             <chr>        
#>  1 08401         16101             CHILLAN      
#>  2 08402         16102             BULNES       
#>  3 08403         16202             COBQUECURA   
#>  4 08404         16203             COELEMU      
#>  5 08405         16302             COIHUECO     
#>  6 08406         16103             CHILLAN VIEJO
#>  7 08407         16104             EL CARMEN    
#>  8 08408         16204             NINHUE       
#>  9 08409         16303             NIQUEN       
#> 10 08410         16105             PEMUCO       
#> # … with 11 more rows