Divisiones electorales oficiales

Contiene las agrupaciones de las distintas comunas en circunscripciones (eleccion de senadores) y distritos (eleccion de diputados) junto con la cantidad de senadores y diputados que le corresponden a cada division segun la legislacion vigente.

divisiones_electorales

Format

Un data frame con 346 observaciones y cinco columnas.

codigo_comuna

vector de texto

codigo_circunscripcion

vector de texto

senadores_circunscripcion

vector numerico

codigo_distrito

vector de texto

diputados_distrito

vector numerico

Source

Elaboracion propia con base en datos de Biblioteca del Congreso Nacional de Chile (BCN).

Examples

divisiones_electorales
#> # A tibble: 346 × 5
#>    codigo_comuna codigo_circunscripcion senadores_circunscripci… codigo_distrito
#>    <chr>         <chr>                                     <int> <chr>          
#>  1 15101         01                                            2 01             
#>  2 15102         01                                            2 01             
#>  3 15202         01                                            2 01             
#>  4 15201         01                                            2 01             
#>  5 01107         02                                            2 02             
#>  6 01402         02                                            2 02             
#>  7 01403         02                                            2 02             
#>  8 01404         02                                            2 02             
#>  9 01101         02                                            2 02             
#> 10 01405         02                                            2 02             
#> # … with 336 more rows, and 1 more variable: diputados_distrito <int>