Mapa de Chile a nivel de circunscripciones

Este mapa es calculado a partir del mapa comunal para no recargar el volumen de datos del paquete.

generar_circunscripciones(mapa = chilemapas::mapa_comunas)

Arguments

mapa
mapa a agregar, por defecto es todo el mapa nacional

Value

Un objeto de clase sf y data.frame.

Examples

generar_circunscripciones()

#> Joining, by = "codigo_circunscripcion"

#> # A tibble: 15 x 3
#>    codigo_circunscrip… codigo_region                                    geometry
#>    <chr>               <chr>                                  <MULTIPOLYGON [°]>
#>  1 02                  01            (((-69.93023 -21.4246, -69.93211 -21.42485…
#>  2 03                  02            (((-70.60654 -23.43054, -70.60175 -23.4351…
#>  3 04                  03            (((-68.36698 -26.79685, -68.36379 -26.8010…
#>  4 05                  04            (((-71.28791 -29.9416, -71.28132 -29.93136…
#>  5 06                  05            (((-71.70008 -33.1555, -71.71085 -33.14618…
#>  6 08                  06            (((-70.91453 -34.14454, -70.91686 -34.1409…
#>  7 09                  07            (((-72.61881 -35.56136, -72.61272 -35.5529…
#>  8 11                  09            (((-73.35579 -38.73982, -73.35826 -38.7262…
#>  9 13                  10            (((-72.99 -41.55301, -72.98086 -41.55318, …
#> 10 14                  11            (((-72.06159 -44.83968, -72.06635 -44.8431…
#> 11 15                  12            (((-72.1423 -54.4747, -72.1399 -54.47724, …
#> 12 07                  13            (((-70.37256 -33.10578, -70.36695 -33.1070…
#> 13 12                  14            (((-73.2668 -39.88605, -73.26215 -39.88671…
#> 14 01                  15            (((-70.35079 -18.8362, -70.34977 -18.83428…
#> 15 10                  16            (((-73.05136 -36.85569, -73.06061 -36.8389…