Mapa de Chile a nivel provincial

Este mapa es calculado a partir del mapa comunal para no recargar el volumen de datos del paquete.

generar_provincias(mapa = chilemapas::mapa_comunas)

Arguments

mapa

mapa a agregar, por defecto es todo el mapa nacional

Value

Un objeto de clase sf y data.frame.

Examples

generar_provincias()
#> # A tibble: 56 × 3
#>    codigo_provincia codigo_region                                       geometry
#>    <chr>            <chr>                                     <MULTIPOLYGON [°]>
#>  1 014              01            (((-68.86081 -21.28512, -68.92172 -21.30035, …
#>  2 011              01            (((-70.09894 -20.08504, -70.01515 -20.07019, …
#>  3 021              02            (((-68.98863 -25.38016, -68.98731 -25.38411, …
#>  4 022              02            (((-68.66131 -23.1077, -68.66418 -23.1052, -6…
#>  5 023              02            (((-69.19173 -22.57654, -69.25181 -22.63043, …
#>  6 032              03            (((-70.68641 -26.15053, -70.68303 -26.15132, …
#>  7 031              03            (((-68.36698 -26.79685, -68.36379 -26.80107, …
#>  8 033              03            (((-70.07472 -28.15016, -70.07486 -28.15708, …
#>  9 043              04            (((-70.59578 -31.27989, -70.60591 -31.28923, …
#> 10 041              04            (((-70.93196 -30.23447, -70.91899 -30.25968, …
#> # … with 46 more rows