Mapa de Chile a nivel de servicios de salud

Este mapa es calculado a partir del mapa comunal para no recargar el volumen de datos del paquete.

generar_servicios_salud(mapa = chilemapas::mapa_comunas)

Arguments

mapa
mapa a agregar, por defecto es todo el mapa nacional

Value

Un objeto de clase sf y data.frame.

Examples

generar_servicios_salud()

#> old-style crs object detected; please recreate object with a recent sf::st_crs()

#> Joining, by = "codigo_comuna"

#> # A tibble: 29 × 3
#>    codigo_servicio_salud codigo_region                                  geometry
#>    <chr>                 <chr>                                <MULTIPOLYGON [°]>
#>  1 02                    01            (((-69.93023 -21.4246, -69.93211 -21.424…
#>  2 03                    02            (((-70.60654 -23.43054, -70.60175 -23.43…
#>  3 04                    03            (((-68.36698 -26.79685, -68.36379 -26.80…
#>  4 05                    04            (((-71.28791 -29.9416, -71.28132 -29.931…
#>  5 06                    05            (((-71.58724 -33.0285, -71.5819 -33.0328…
#>  6 07                    05            (((-71.55077 -32.94576, -71.5534 -32.939…
#>  7 08                    05            (((-70.1475 -32.72665, -70.13743 -32.731…
#>  8 15                    06            (((-70.91453 -34.14454, -70.91686 -34.14…
#>  9 16                    07            (((-72.61881 -35.56136, -72.61272 -35.55…
#> 10 18                    08            (((-73.05136 -36.85569, -73.06061 -36.83…
#> # … with 19 more rows