Mapa de Chile a nivel zonal

Contiene todas las zonas urbanas del territorio nacional indicando las divisiones administrativas (comuna, provincia y region) a las que pertenecen.

mapa_zonas

Format

Un data frame con 4852 observaciones y cinco columnas.

geocodigo

vector de texto

codigo_comuna

vector de texto

codigo_provincia

vector de texto

codigo_region

vector de texto

geometry

vector de poligonos

Source

Elaboracion propia con base en datos de Instituto Nacional de Estadisticas (INE) y Subsecretaria de Desarrollo Regional (SUBDERE).

Examples

mapa_zonas

#> Simple feature collection with 4852 features and 4 fields
#> geometry type:  MULTIPOLYGON
#> dimension:      XY
#> bbox:           xmin: -109.4428 ymin: -54.93854 xmax: -67.59017 ymax: -18.19346
#> epsg (SRID):    NA
#> proj4string:    +proj=longlat +ellps=GRS80 +no_defs
#> First 10 features:
#>     geocodigo codigo_comuna codigo_provincia codigo_region
#> 1  1405011001          1405              014            01
#> 2  1401011001          1401              014            01
#> 3  1401011002          1401              014            01
#> 4  1404011001          1404              014            01
#> 5  1107011002          1107              011            01
#> 6  1107031002          1107              011            01
#> 7  1107011003          1107              011            01
#> 8  1107021004          1107              011            01
#> 9  1107021001          1107              011            01
#> 10 1107021002          1107              011            01
#>                          geometry
#> 1  MULTIPOLYGON (((-69.31192 -...
#> 2  MULTIPOLYGON (((-69.78676 -...
#> 3  MULTIPOLYGON (((-69.78078 -...
#> 4  MULTIPOLYGON (((-69.7696 -1...
#> 5  MULTIPOLYGON (((-70.06361 -...
#> 6  MULTIPOLYGON (((-70.08164 -...
#> 7  MULTIPOLYGON (((-70.09046 -...
#> 8  MULTIPOLYGON (((-70.09813 -...
#> 9  MULTIPOLYGON (((-70.10061 -...
#> 10 MULTIPOLYGON (((-70.10293 -...