Mapa de Chile a nivel comunal

Contiene todas las comunas del territorio nacional indicando las divisiones administrativas (provincia y region) a las que pertenecen.

mapa_comunas

Format

Un data frame con 346 observaciones y cuatro columnas.

codigo_comuna

vector de texto

codigo_provincia

vector de texto

codigo_region

vector de texto

geometry

vector de poligonos

Source

Elaboracion propia con base en datos de Instituto Nacional de Estadisticas (INE) y Subsecretaria de Desarrollo Regional (SUBDERE).

Examples

mapa_comunas

#> # A tibble: 345 x 4
#>    codigo_comuna codigo_provincia codigo_region                         geometry
#>    <chr>         <chr>            <chr>                       <MULTIPOLYGON [°]>
#>  1 01401         014              01            (((-68.86081 -21.28512, -68.921…
#>  2 01403         014              01            (((-68.65113 -19.77188, -68.811…
#>  3 01405         014              01            (((-68.65113 -19.77188, -68.635…
#>  4 01402         014              01            (((-69.31789 -19.13651, -69.271…
#>  5 01404         014              01            (((-69.39615 -19.06125, -69.400…
#>  6 01107         011              01            (((-70.1095 -20.35131, -70.1243…
#>  7 01101         011              01            (((-70.09894 -20.08504, -70.102…
#>  8 02104         021              02            (((-68.98863 -25.38016, -68.987…
#>  9 02101         021              02            (((-70.60654 -23.43054, -70.601…
#> 10 02201         022              02            (((-67.94302 -22.38175, -67.955…
#> # … with 335 more rows